requirements.txt
asgiref==3.7.2
cloudinary==1.34.0
dj-database-url==2.1.0
dj-rest-auth==2.1.9
dj3-cloudinary-storage==0.0.6
Django==3.2.20
django-allauth==0.44.0
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
oauthlib==3.2.2
Pillow==10.0.0
PyJWT==2.8.0
python3-openid==3.2.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
urllib3==1.26.16
